Axis settings now double precision, bug in parsing input values!

+++ java/pro2/src/pro2/Plot/RectPlot/RectPlotSettings.java
@@ -6,12 +6,12 @@ public class RectPlotSettings {
     // Public Data
     //================================================================================
 
-	public int xAxisMinimum = 0;
-	public int xAxisMaximum = 0;
+	public double xAxisMinimum = 0;
+	public double xAxisMaximum = 0;
 	public int xAxisSteps = 0;
 	
-	public int yAxisMinimum = 0;
-	public int yAxisMaximum = 0;
+	public double yAxisMinimum = 0;
+	public double yAxisMaximum = 0;
 	public int yAxisSteps = 0;
 	
 	//================================================================================
